APIs/watchful.li/1.0.0/python-aiohttp-client/tests/test_feedbacks_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.audit import Audit
from openapi_server.models.feedback import Feedback


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("*/* not supported by Connexion. Use application/json instead. See https://github.com/zalando/connexion/pull/760")
async def test_create_feedbacks(client):
    """Test case for create_feedbacks

    Create a feedback
    """
    body = openapi_server.Feedback()
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='POST',
        path='/api/v1/feedbacks',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_feedbacks(client):
    """Test case for get_feedbacks

    Get feedbacks
    """
    params = [('fields', 'fields_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/v1/feedbacks',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_fields_feedbacks(client):
    """Test case for get_fields_feedbacks

    Get the list of fields
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/v1/feedbacks/metadata',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

